+++ lib/features/dashboards/ui/widgets/charts/dashboard_workout_chart.dart
@@ -5,13 +5,13 @@ import 'package:flutter/material.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
+import 'package:lotti/features/dashboards/state/workout_data.dart';
 import 'package:lotti/features/dashboards/ui/widgets/charts/dashboard_chart.dart';
+import 'package:lotti/features/dashboards/ui/widgets/charts/time_series/time_series_bar_chart.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/health_import.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/color.dart';
-import 'package:lotti/widgets/charts/dashboard_workout_data.dart';
-import 'package:lotti/widgets/charts/time_series/time_series_bar_chart.dart';
 import 'package:lotti/widgets/charts/utils.dart';
 
 class DashboardWorkoutChart extends StatefulWidget {
