+++ lib/features/dashboards/ui/widgets/charts/time_series/time_series_line_chart.dart
@@ -1,25 +1,19 @@
 import 'dart:core';
-import 'dart:math';
 
-import 'package:collection/collection.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
+import 'package:lotti/features/dashboards/ui/widgets/charts/time_series/utils.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/utils/date_utils_extension.dart';
 import 'package:lotti/utils/platform.dart';
-import 'package:lotti/widgets/charts/time_series/utils.dart';
 import 'package:lotti/widgets/charts/utils.dart';
 import 'package:tinycolor2/tinycolor2.dart';
 
-class TimeSeriesBarChart extends StatelessWidget {
-  const TimeSeriesBarChart({
+class TimeSeriesLineChart extends StatelessWidget {
+  const TimeSeriesLineChart({
     required this.data,
     required this.rangeStart,
     required this.rangeEnd,
-    required this.colorByValue,
     this.unit = '',
-    this.valueInHours = false,
     super.key,
   });
 
@@ -27,24 +21,9 @@ class TimeSeriesBarChart extends StatelessWidget {
   final DateTime rangeStart;
   final DateTime rangeEnd;
   final String unit;
-  final bool valueInHours;
-  final ColorByValue colorByValue;
 
   @override
   Widget build(BuildContext context) {
-    final inRange = daysInRange(rangeStart: rangeStart, rangeEnd: rangeEnd);
-
-    final byDay = <String, Observation>{};
-    for (final observation in data) {
-      final day = observation.dateTime.ymd;
-      byDay[day] = observation;
-    }
-
-    final dataWithEmptyDays = inRange.map((day) {
-      final observation = byDay[day] ?? Observation(DateTime.parse(day), 0);
-      return observation;
-    });
-
     final rangeInDays = rangeEnd.difference(rangeStart).inDays;
 
     final gridInterval = rangeInDays > 182
@@ -55,36 +34,19 @@ class TimeSeriesBarChart extends StatelessWidget {
                 ? 7
                 : 1;
 
-    final screenWidth = MediaQuery.sizeOf(context).width;
-    final barsWidth =
-        (screenWidth - 150 - rangeInDays - screenWidth * 0.1) / rangeInDays;
-
-    final barGroups = dataWithEmptyDays
-        .sortedBy((observation) => observation.dateTime)
-        .map((observation) {
-      return BarChartGroupData(
-        x: observation.dateTime.millisecondsSinceEpoch,
-        barRods: [
-          BarChartRodData(
-            toY: observation.value.toDouble(),
-            borderRadius: const BorderRadius.only(
-              topLeft: Radius.circular(1),
-              topRight: Radius.circular(1),
-            ),
-            color: colorByValue(observation),
-            width: max(barsWidth, 1),
+    final spots = data
+        .map(
+          (item) => FlSpot(
+            item.dateTime.millisecondsSinceEpoch.toDouble(),
+            item.value.toDouble(),
           ),
-        ],
-      );
-    }).toList();
+        )
+        .toList();
 
-    Widget bottomTitleWidgets(
-      double value,
-      TitleMeta meta,
-    ) {
+    Widget bottomTitleWidgets(double value, TitleMeta meta) {
       final ymd = DateTime.fromMillisecondsSinceEpoch(value.toInt());
       if (ymd.day == 1 ||
-          (rangeInDays < 92 && ymd.day == 15) ||
+          (rangeInDays < 90 && ymd.day == 15) ||
           (rangeInDays < 30 && ymd.day == 8) ||
           (rangeInDays < 30 && ymd.day == 22)) {
         return SideTitleWidget(
@@ -100,9 +62,8 @@ class TimeSeriesBarChart extends StatelessWidget {
         top: 20,
         right: 20,
       ),
-      child: BarChart(
-        BarChartData(
-          groupsSpace: 5,
+      child: LineChart(
+        LineChartData(
           gridData: FlGridData(
             show: false,
             horizontalInterval: double.maxFinite,
@@ -111,8 +72,9 @@ class TimeSeriesBarChart extends StatelessWidget {
             getDrawingHorizontalLine: (value) => gridLine,
             getDrawingVerticalLine: (value) => gridLine,
           ),
-          barTouchData: BarTouchData(
-            touchTooltipData: BarTouchTooltipData(
+          clipData: const FlClipData.horizontal(),
+          lineTouchData: LineTouchData(
+            touchTooltipData: LineTouchTooltipData(
               tooltipMargin: isMobile ? 24 : 16,
               tooltipPadding: const EdgeInsets.symmetric(
                 horizontal: 8,
@@ -121,17 +83,27 @@ class TimeSeriesBarChart extends StatelessWidget {
               getTooltipColor: (_) =>
                   Theme.of(context).primaryColor.desaturate(),
               tooltipRoundedRadius: 8,
-              getTooltipItem: (groupData, timestamp, rodData, foo) {
-                final formatted = valueInHours
-                    ? hoursToHhMm(rodData.toY)
-                    : NumberFormat('#,###').format(rodData.toY);
-                return BarTooltipItem(
-                  '$formatted $unit\n'
-                  '${chartDateFormatterYMD(groupData.x)}',
-                  chartTooltipStyleBold.copyWith(
-                    color: context.colorScheme.onPrimary,
-                  ),
-                );
+              getTooltipItems: (List<LineBarSpot> spots) {
+                return spots.map((spot) {
+                  return LineTooltipItem(
+                    '',
+                    TextStyle(
+                      fontSize: fontSizeSmall,
+                      fontWeight: FontWeight.w300,
+                      color: context.colorScheme.onPrimary,
+                    ),
+                    children: [
+                      TextSpan(
+                        text: '${spot.y.toInt()} $unit\n',
+                        style: chartTooltipStyleBold,
+                      ),
+                      TextSpan(
+                        text: chartDateFormatterFull(spot.x),
+                        style: chartTooltipStyle,
+                      ),
+                    ],
+                  );
+                }).toList();
               },
             ),
           ),
@@ -158,9 +130,29 @@ class TimeSeriesBarChart extends StatelessWidget {
             show: true,
             border: Border.all(color: const Color(0xff37434d)),
           ),
-          barGroups: barGroups,
+          minX: rangeStart.millisecondsSinceEpoch.toDouble(),
+          maxX: rangeEnd.millisecondsSinceEpoch.toDouble(),
+          lineBarsData: [
+            LineChartBarData(
+              spots: spots,
+              gradient: LinearGradient(
+                colors: gradientColors,
+              ),
+              dotData: const FlDotData(
+                show: false,
+              ),
+              belowBarData: BarAreaData(
+                show: true,
+                gradient: LinearGradient(
+                  colors: gradientColors
+                      .map((color) => color.withAlpha(77))
+                      .toList(),
+                ),
+              ),
+            ),
+          ],
         ),
-        //duration: Duration.zero,
+        duration: Duration.zero,
       ),
     );
   }
