+++ lib/features/dashboards/ui/widgets/charts/time_series/time_series_multiline_chart.dart
@@ -1,29 +1,36 @@
 import 'dart:core';
+import 'dart:math';
 
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
+import 'package:lotti/features/dashboards/ui/widgets/charts/time_series/utils.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/platform.dart';
-import 'package:lotti/widgets/charts/time_series/utils.dart';
 import 'package:lotti/widgets/charts/utils.dart';
 import 'package:tinycolor2/tinycolor2.dart';
 
-class TimeSeriesLineChart extends StatelessWidget {
-  const TimeSeriesLineChart({
-    required this.data,
+class TimeSeriesMultiLineChart extends StatelessWidget {
+  const TimeSeriesMultiLineChart({
+    required this.lineBarsData,
     required this.rangeStart,
     required this.rangeEnd,
+    required this.minVal,
+    required this.maxVal,
     this.unit = '',
     super.key,
   });
 
-  final List<Observation> data;
+  final List<LineChartBarData> lineBarsData;
   final DateTime rangeStart;
   final DateTime rangeEnd;
+  final num minVal;
+  final num maxVal;
   final String unit;
 
   @override
   Widget build(BuildContext context) {
+    final valRange = maxVal - minVal;
+
     final rangeInDays = rangeEnd.difference(rangeStart).inDays;
 
     final gridInterval = rangeInDays > 182
@@ -33,15 +40,6 @@ class TimeSeriesLineChart extends StatelessWidget {
             : rangeInDays > 30
                 ? 7
                 : 1;
-
-    final spots = data
-        .map(
-          (item) => FlSpot(
-            item.dateTime.millisecondsSinceEpoch.toDouble(),
-            item.value.toDouble(),
-          ),
-        )
-        .toList();
 
     Widget bottomTitleWidgets(double value, TitleMeta meta) {
       final ymd = DateTime.fromMillisecondsSinceEpoch(value.toInt());
@@ -72,7 +70,6 @@ class TimeSeriesLineChart extends StatelessWidget {
             getDrawingHorizontalLine: (value) => gridLine,
             getDrawingVerticalLine: (value) => gridLine,
           ),
-          clipData: const FlClipData.horizontal(),
           lineTouchData: LineTouchData(
             touchTooltipData: LineTouchTooltipData(
               tooltipMargin: isMobile ? 24 : 16,
@@ -121,6 +118,7 @@ class TimeSeriesLineChart extends StatelessWidget {
             leftTitles: const AxisTitles(
               sideTitles: SideTitles(
                 showTitles: true,
+                interval: double.maxFinite,
                 getTitlesWidget: leftTitleWidgets,
                 reservedSize: 40,
               ),
@@ -132,25 +130,9 @@ class TimeSeriesLineChart extends StatelessWidget {
           ),
           minX: rangeStart.millisecondsSinceEpoch.toDouble(),
           maxX: rangeEnd.millisecondsSinceEpoch.toDouble(),
-          lineBarsData: [
-            LineChartBarData(
-              spots: spots,
-              gradient: LinearGradient(
-                colors: gradientColors,
-              ),
-              dotData: const FlDotData(
-                show: false,
-              ),
-              belowBarData: BarAreaData(
-                show: true,
-                gradient: LinearGradient(
-                  colors: gradientColors
-                      .map((color) => color.withAlpha(77))
-                      .toList(),
-                ),
-              ),
-            ),
-          ],
+          minY: max(minVal - valRange * 0.2, 0),
+          maxY: maxVal + valRange * 0.2,
+          lineBarsData: lineBarsData,
         ),
         duration: Duration.zero,
       ),
